Added artifact uploading functionality
With this came another cloud exception to raise where an artifact
with same name exists. Added a uploading test. Also changed
requirements to include filechunkio.

+++ pyshelf/cloud/storage.py
@@ -1,8 +1,10 @@
 from boto.s3.connection import S3Connection
 from boto.s3.key import Key
 import os
+import math
+from filechunkio import FileChunkIO
 from pyshelf.cloud.stream_iterator import StreamIterator
-from pyshelf.cloud.cloud_exceptions import ArtifactNotFoundError, BucketNotFoundError
+from pyshelf.cloud.cloud_exceptions import ArtifactNotFoundError, BucketNotFoundError, DuplicateArtifactError
 
 
 class Storage(object):
@@ -44,16 +46,29 @@ class Storage(object):
         stream = StreamIterator(key)
         return stream
 
-    def upload_artifact(self, artifact_name, fp):
+    def upload_artifact(self, artifact_path, artifact_name, src):
         bucket = self._get_bucket(self.bucket_name)
-        key = Key(bucket, artifact_name)
-        key.set_contents_from_file(fp)
-        return True
+        if bucket.get_key(artifact_name) is not None:
+            raise DuplicateArtifactError(artifact_name)
+        src_size = os.stat(src).st_size
+        """ arbitrarily chunked at 100 MB """
+        chunk_size = 104857600
+        chunk_count = int(math.ceil(src_size/float(chunk_size)))
+        self.logger.debug("Initiating upload")
+        mp = bucket.initiate_multipart_upload(artifact_name)
+        for i in range(chunk_count):
+            offset = chunk_size * i
+            bytes = min(chunk_size, src_size - offset)
+            with FileChunkIO(src, 'r', offset=offset, bytes=bytes) as fp:
+                mp.upload_part_from_file(fp, part_num=i + 1)
+        mp.complete_upload()
+        self.logger.debug("Completed upload of {}".format(artifact_name))
 
     def delete_artifact(self, artifact_name):
         key = self._get_key(artifact_name)
+        if key is None:
+            raise ArtifactNotFoundError(artifact_name)
         key.delete()
-        return True
 
     def _get_key(self, artifact_name):
         bucket = self._get_bucket(self.bucket_name)
